Add files via upload
Concepts of computer vision using OpenCV library was used. Three channel image was converted into 1 channel (gray scale) image. Gaussian Blur was applied to smoothen and reduce the noise of image. Then canny function was used to identify the edges. Region of interest was developed in the image using which Hough Transformation was applied to detect the straight line.